+++ src/server/domain/entity/base/index.ts
