+++ src/server/domain/entity/error/index.ts
